Add type definitions to FMI 2

+++ rust/fmi-rs/fmi-rs/src/model_description/fmi3.rs
@@ -27,7 +27,7 @@ impl FromStr for VariableNamingConvention {
 #[derive(Debug)]
 pub struct Item {
     pub name: String,
-    pub value: i64,
+    pub value: i32,
     pub description: Option<String>,
 }
